[DDX-3] splitted activity and data model and update some functions
